use Jac points in Lagrange basis transformation
